250403 feat/fix - changed DataSource & Service Structure for safe udp transmit

+++ nexus/SourceData/src/main/java/com/hooniegit/SourceData/Source/Specified.java
@@ -6,16 +6,15 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 /**
- * Body 타입으로 사용할 데이터 클래스입니다.<br>
- * - 해당 클래스는 테스트용으로 현장에서 사용하지 않습니다.<br>
- * - id 단위 데이터와 category 단위 데이터로 구분됩니다.
+ * Body 타입으로 사용할 데이터 클래스입니다.
+ *  * 해당 클래스는 테스트용으로 현장에서 사용하지 않습니다.
  */
 
 @Getter @Setter @AllArgsConstructor @NoArgsConstructor
 public class Specified {
 
     private int id;
-    private long value;
+    private double value;
     private boolean state;
 
     private int group;
